requirements.txt
clickhouse-connect==0.7.8
minio==7.2.7
python-dotenv==1.0.1
# pyspark==3.5.1